# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLane API
  version: 0.1.0
  description: >
    The OpenLane public API — source of truth (ADR-0005).
    Every UI action hits this API. Versioned under /v1; breaking changes need /v2.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080/v1
    description: local
paths:
...
  /auth/magic-link:
    post:
      summary: Request a magic-link login token (email + workspace slug)
      description: >
        Always returns 202 (never reveal whether the email exists). In dev
        mode (OPENLANE_DEV_LOGIN=1) the response includes dev_token so
        flows work without SMTP. Token is single-use, 15 minutes.
      operationId: requestMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, workspace_slug]
              properties:
                email: {type: string, format: email}
                workspace_slug: {type: string}
      responses:
        "202":
          description: Token sent (dev_token included when dev login enabled)
          content:
            application/json:
              schema:
                type: object
                properties:
                  dev_token: {type: string}
        "400":
          description: Malformed request
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/magic-link/consume:
    get:
      summary: Exchange a magic-link token for session tokens
      operationId: consumeMagicLink
      parameters:
        - name: token
          in: query
          required: true
          schema: {type: string, minLength: 43}
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionTokens"
        "401":
          description: Invalid, expired, or already-used token
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
  /auth/refresh:
    post:
      summary: Rotate a refresh token (family theft detection)
      description: >
        Refresh tokens rotate on every use. Presenting an already-rotated
        token revokes the entire family (assumed theft).
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string, minLength: 43}
      responses:
        "200":
          description: New session tokens
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionTokens"}
        "401":
          description: Unknown or revoked token (family revoked if reuse)
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
...
components:
...
  schemas:
    SessionTokens:
      type: object
      properties:
        access_token: {type: string, description: "JWT, 15 min"}
        refresh_token: {type: string, minLength: 43}
        token_type: {type: string, const: Bearer}
        user_id: {type: string, format: uuid}
        workspace_id: {type: string, format: uuid}
        role: {type: string}
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: RFC 9457 problem type URI
          examples: [https://openlane.dev/errors/portal-token-expired]
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    portalToken:
      type: apiKey
      in: query
      name: portal_token
security:
  - bearerAuth: []